+++ NEWS.md
@@ -8,6 +8,8 @@ Version 1.0.0b0 - dev
 - Release notes: https://ezdxf.mozman.at/release-v1-0.html
 - NEW: `Drawing.paperspace()` type-safe method to acquire paperspace layouts 
 - CHANGE: removed deprecated features
+- BUGFIX: [#747](https://github.com/mozman/ezdxf/issues/747)
+  fix virtual entities of 3D DIMENSION entities  
 
 Version 0.18.1 - 2022-09-03
 ---------------------------